- Added RFC 8976 - Added Zone Validation - Optimized Zone Signing - Optimized Zone masterfile parsing - Optimized pipelined client transport

+++ ARSoft.Tools.Net/Dns/DnsSec/NSec3ParamRecord.cs
@@ -96,7 +96,7 @@ namespace ARSoft.Tools.Net.Dns
 		internal override string RecordDataToString()
 		{
 			return (byte) HashAlgorithm
-			       + " " + Flags
+			       + " " + (byte)Flags
 			       + " " + Iterations
 			       + " " + ((Salt.Length == 0) ? "-" : Salt.ToBase16String());
 		}
